修改了log.c和log.h文件：增加了日志缓冲区大小，并更新了日志回调类型定义。

+++ log.c
@@ -10,7 +10,7 @@ void log_init(log_callback_t log)
 void log_message(int level, const char *fmt, ...)
 {
     char buffer[512];
-    int len = 0;
+    uint16_t len = 0;
 
     switch (level) {
         case LOG_DEBUG:
